C header errors
