+++ source/generic/mtree_text.c
@@ -10,6 +10,10 @@
 #include "mtree_gist.h"
 #include "mtree_utils_var.h"
 
+#ifdef PG_MODULE_MAGIC
+PG_MODULE_MAGIC;
+#endif
+
 #define SameStrategyNumber      1
 #define OverlapStrategyNumber   2
 #define ContainsStrategyNumber  3
@@ -38,7 +42,7 @@ typedef enum {
 } PicksplitStrategy;
 
 UnionStrategy unionStrategy = Best;
-PicksplitStrategy picksplitStrategy = Random;
+PicksplitStrategy picksplitStrategy = SamplingMinCoveringMax;
 
 PG_FUNCTION_INFO_V1(gmt_text_compress);
 PG_FUNCTION_INFO_V1(gmt_bpchar_compress);
@@ -48,6 +52,110 @@ PG_FUNCTION_INFO_V1(gmt_text_consistent);
 PG_FUNCTION_INFO_V1(gmt_bpchar_consistent);
 PG_FUNCTION_INFO_V1(gmt_text_penalty);
 PG_FUNCTION_INFO_V1(gmt_text_same);
+
+PG_FUNCTION_INFO_V1(gmt_text_distance);
+PG_FUNCTION_INFO_V1(gmt_text_overlap);
+PG_FUNCTION_INFO_V1(gmt_text_contains);
+PG_FUNCTION_INFO_V1(gmt_text_contained);
+PG_FUNCTION_INFO_V1(gmt_text_distance_internal);
+
+PG_FUNCTION_INFO_V1(gmt_decompress);
+PG_FUNCTION_INFO_V1(gmtreekey_in);
+PG_FUNCTION_INFO_V1(gmtreekey_out);
+
+/**************************************************
+ * In/Out for keys
+ **************************************************/
+
+Datum gmtreekey_in(PG_FUNCTION_ARGS)
+{
+  ereport(ERROR,
+          (errcode(ERRCODE_FEATURE_NOT_SUPPORTED),
+           errmsg("<datatype>key_in() not implemented")));
+
+  PG_RETURN_POINTER(NULL);
+}
+
+
+
+Datum gmtreekey_out(PG_FUNCTION_ARGS)
+{
+  ereport(ERROR,
+          (errcode(ERRCODE_FEATURE_NOT_SUPPORTED),
+           errmsg("<datatype>key_out() not implemented")));
+
+  PG_RETURN_POINTER(NULL);
+}
+
+/* GiST DeCompress methods do not do anything. */
+Datum gmt_decompress(PG_FUNCTION_ARGS)
+{
+  PG_RETURN_POINTER(PG_GETARG_POINTER(0));
+}
+
+
+/* Used for key sorting */
+typedef struct
+{
+  int i;
+  GMT_VARKEY *t;
+} Vsrt;
+
+typedef struct
+{
+  const gmtree_vinfo *tinfo;
+  Oid collation;
+  FmgrInfo *flinfo;
+} gmt_vsrt_arg;
+
+PG_FUNCTION_INFO_V1(gmt_var_decompress);
+PG_FUNCTION_INFO_V1(gmt_var_fetch);
+
+Datum gmt_var_decompress(PG_FUNCTION_ARGS)
+{
+  GISTENTRY *entry = (GISTENTRY *)PG_GETARG_POINTER(0);
+  GMT_VARKEY *key = (GMT_VARKEY *)PG_DETOAST_DATUM(entry->key);
+
+  if (key != (GMT_VARKEY *)DatumGetPointer(entry->key))
+  {
+    GISTENTRY *retval = (GISTENTRY *)palloc(sizeof(GISTENTRY));
+
+    gistentryinit(*retval, PointerGetDatum(key), entry->rel, entry->page, entry->offset, false);
+
+    PG_RETURN_POINTER(retval);
+  }
+
+  PG_RETURN_POINTER(entry);
+}
+
+/* Returns a better readable representation of variable key ( sets pointer ) */
+GMT_VARKEY_R gmt_var_key_readable(const GMT_VARKEY *k)
+{
+  GMT_VARKEY_R r;
+  r.lower = (bytea *)&(((char *)k)[VARHDRSZ]);
+  if (VARSIZE(k) > (VARHDRSZ + (VARSIZE(r.lower))))
+  {
+    r.upper = (bytea *)&(((char *)k)[VARHDRSZ + INTALIGN(VARSIZE(r.lower))]);
+  }
+  else
+  {
+    r.upper = r.lower;
+  }
+  return r;
+}
+
+Datum gmt_var_fetch(PG_FUNCTION_ARGS)
+{
+  GISTENTRY *entry = (GISTENTRY *)PG_GETARG_POINTER(0);
+  GMT_VARKEY *key = (GMT_VARKEY *)PG_DETOAST_DATUM(entry->key);
+  GMT_VARKEY_R r = gmt_var_key_readable(key);
+  GISTENTRY *retval;
+
+  retval = palloc(sizeof(GISTENTRY));
+  gistentryinit(*retval, PointerGetDatum(r.lower), entry->rel, entry->page, entry->offset, true);
+
+  PG_RETURN_POINTER(retval);
+}
 
 /**************************************************
  * For comparison
@@ -119,6 +227,32 @@ static gmtree_vinfo tinfo =
 /**************************************************
  * Useful functions
  **************************************************/
+
+Datum gmt_text_distance(PG_FUNCTION_ARGS) {
+  int asd = 1;
+  ereport(ERROR, (errcode(ERRCODE_WARNING_DYNAMIC_RESULT_SETS_RETURNED), errmsg("gmt_text_distance")));
+  PG_RETURN_POINTER(string_distance(PG_GETARG_POINTER(0), PG_GETARG_POINTER(1)));
+}
+
+Datum gmt_text_overlap(PG_FUNCTION_ARGS) {
+  ereport(ERROR, (errcode(ERRCODE_WARNING_DYNAMIC_RESULT_SETS_RETURNED), errmsg("gmt_text_overlap")));
+  PG_RETURN_POINTER(NULL);
+}
+
+Datum gmt_text_contains(PG_FUNCTION_ARGS) {
+  ereport(ERROR, (errcode(ERRCODE_WARNING_DYNAMIC_RESULT_SETS_RETURNED), errmsg("gmt_text_contains")));
+  PG_RETURN_POINTER(NULL);
+}
+
+Datum gmt_text_contained(PG_FUNCTION_ARGS) {
+  ereport(ERROR, (errcode(ERRCODE_WARNING_DYNAMIC_RESULT_SETS_RETURNED), errmsg("gmt_text_contained")));
+  PG_RETURN_POINTER(NULL);
+}
+
+Datum gmt_text_distance_internal(PG_FUNCTION_ARGS) {
+  ereport(ERROR, (errcode(ERRCODE_WARNING_DYNAMIC_RESULT_SETS_RETURNED), errmsg("gmt_text_distance_internal")));
+  PG_RETURN_POINTER(NULL);
+}
 
 int string_distance(const char* a, const char* b)
 {
